--- a/tapdb/sqlc/migrations/000024_universe_optimization_indexes_queries.down.sql
+++ b/tapdb/sqlc/migrations/000024_universe_optimization_indexes_queries.down.sql
@@ -1,0 +1,5 @@
+DROP INDEX IF EXISTS idx_mssmt_nodes_stats;
+
+DROP INDEX IF EXISTS idx_universe_leaves_composite;
+
+DROP INDEX IF EXISTS idx_universe_roots_composite;