--- a/tapdb/universe_perf_short_test.go
+++ b/tapdb/universe_perf_short_test.go
@@ -1,0 +1,9 @@
+//go:build !longtests
+
+package tapdb
+
+var (
+	numAssets        = 100
+	numLeavesPerTree = 300
+	numQueries       = 100
+)