# Universe Query Performance Optimization and Lock Safety

This PR introduces two key improvements:
1. Strategic composite indices to optimize `tapdb` universe queries
2. Enhanced lock safety for concurrent stats access

## Lock Safety Improvements
```go
// Added finer-grained locking for stats access
statsCacheLogger    *cacheLogger
statsRefresh        *time.Timer
eventsMtx          sync.RWMutex    // Upgraded from sync.Mutex
assetEventsCache   assetEventsCache
eventsCacheLogger  *cacheLogger
syncStatsMtx      sync.RWMutex    // Upgraded from sync.Mutex
syncStatsCache    *atomicSyncStatsCache
syncStatsRefresh  *time.Timer
```

Key changes:
- Upgraded mutexes to RWMutex for better concurrency
- Separated read/write lock paths in QueryAssetStatsPerDay and QuerySyncStats
- Proper lock ordering to prevent deadlocks
- Added atomic cache for sync stats

## Performance Optimizations: Strategic Indexing

### Key Changes
```sql
-- Optimizes universe root lookups and filtering
CREATE INDEX idx_universe_roots_composite
ON universe_roots(namespace_root, proof_type, asset_id);

-- Enhances universe leaf queries and joins
CREATE INDEX idx_universe_leaves_composite
ON universe_leaves(leaf_node_namespace, universe_root_id, leaf_node_key);

-- Improves MSSMT node lookups and aggregations 
CREATE INDEX idx_mssmt_nodes_composite 
ON mssmt_nodes(namespace, key, hash_key, sum);
```

### Performance Impact

#### PostgreSQL 
- Asset Stats Query: 4.58x improvement (205ms → 44.8ms)
  - Query cost reduced from 49.17 to 23.58
  - Eliminated sequence scans
  - Optimized sort operations and hash joins

- Universe Leaves Query: 1.24x improvement (20.9ms → 16.8ms)
  - Query cost reduced from 59.91 to 20.99
  - Replaced sequence scans with efficient index usage

- Root Query: 1.05x improvement (17.7ms → 16.9ms)
  - Maintained efficient lookups with better join patterns

#### SQLite
- Asset Stats Query: 2.02x improvement (55.9ms → 27.7ms)
  - Replaced BLOOM FILTER with covering index
  - More efficient join strategy

- Universe Leaves and Root Queries: Maintain baseline performance
  - Slight overhead (~5%) from index maintenance
  - Already optimized by SQLite's automatic indexing

#### Write Performance
- Single inserts: <5% overhead
- Batch inserts (100+): 10-15% overhead
- Storage impact: ~25% additional space across all indices

### Technical Details

#### Query Cost Analysis
PostgreSQL shows significant cost reductions:
- Complex aggregations: 52% reduction in total cost
- Join operations: 65% reduction in materialization overhead
- Sort operations: ~50% reduction in cost

The improvements come from:
- Better join order selection through index availability
- Reduced materialization steps
- Elimination of table scans
- More efficient sort operations through index coverage

### Implementation Considerations
- Indices designed to balance read/write performance
- Column ordering optimized for common query patterns
- Careful consideration of storage overhead vs performance gains
- Support for both PostgreSQL and SQLite query planners

## Testing
- Comprehensive benchmarks across both databases
- Test dataset: 100 assets, 300 leaves per tree
- 100 iterations per query type
- Full query plan analysis for optimization verification

## Next Steps
1. Monitor index usage patterns in production
2. Gather metrics on maintenance overhead
3. Consider database-specific optimizations
4. Evaluate impact on write-heavy workflows
5. Monitor lock contention in production

The PR delivers two key improvements: significantly better query performance through strategic indexing and enhanced concurrency safety through improved locking patterns. Both changes maintain reasonable overhead while providing substantial benefits for production deployments.