@jbrill 2024-11-05
note: incorporating postgres functionality now for the test