@jbrill 2024-11-12
@guggero thanks so much for the feedback. I've made those changes, which include some sweeping changes to both the indices and the performance test.

I opted to use the `EXPLAIN ANALYZE` statement to get more insightful details of our indices and how they are altering our main queries.

Here is the output for [postgres](https://pastebin.com/dRfQP5J4) and [sqlite](https://pastebin.com/VrRL73mf).