@guggero 2024-11-13
Can you please squash the two middle commits so it's a bit easier to review? Going to take another look next week.