@jbrill 2024-11-13
> Can you please squash the two middle commits so it's a bit easier to review? Going to take another look next week.

Sounds good. I've also updated the PR description to reflect the lock changes and analysis from the query plans.