@jbrill 2024-11-20
@guggero thanks for the round of feedback! I've patched that commit onto the performance test commit in this PR. I've also updated the code style based on the formatting rules doc. Thanks a bunch for this reference.